requirements.txt
psycopg2==2.9.9
django-ckeditor-5==0.2.12
Unidecode==1.3.8
django-mptt==0.16.0
django-mptt-admin==2.6.2